Vulkan: Texture - Added Vulkan texture base
- Added Vulkan texture base
- Fixed anisotropy type from int to float

+++ Core/src/Texture/TextureBase.h
@@ -27,7 +27,7 @@ namespace Texture
         TextureSamplerAddressingMode addressingModeU;
         TextureSamplerAddressingMode addressingModeV;
         TextureSamplerAddressingMode addressingModeW;
-        UINT                         anisotropyNum;
+        float                        anisotropyNum;
         Math::Vector4                borderColor;
         BOOL                         normalize;
         TextureSamplerFilter         mipmapFilter;
